emails/src/requirements.txt
azure-functions
azure-cli
azure-storage-blob
azure-keyvault-secrets
azure-identity
python-dotenv
requests
